backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
requests-oauthlib==2.0.0
starlette==0.41.3
itsdangerous==2.2.0
